+++ asr-service/app/api/compat/mappers.py
@@ -6,41 +6,9 @@
 """
 from __future__ import annotations
 
-# ─── 语言码归一（上游 ISO-639-1 → Qwen3-ASR 规范名）───
-#
-# OpenAI/DashScope 客户端按 ISO-639-1 传 language（zh/en/…），但 Qwen3-ASR 引擎只认
-# 规范英文名（Chinese/English/…，见 qwen_asr.validate_language——它把 zh 归一成 Zh 再校验，
-# 必挂）。此处一次性翻译；未识别的码降级为 None（交模型自动检测），不让非法 hint 击穿到引擎抛错。
-_ISO_TO_QWEN = {
-    "zh": "Chinese", "en": "English", "yue": "Cantonese", "ar": "Arabic",
-    "de": "German", "fr": "French", "es": "Spanish", "pt": "Portuguese",
-    "id": "Indonesian", "it": "Italian", "ko": "Korean", "ru": "Russian",
-    "th": "Thai", "vi": "Vietnamese", "ja": "Japanese", "tr": "Turkish",
-    "hi": "Hindi", "ms": "Malay", "nl": "Dutch", "sv": "Swedish",
-    "da": "Danish", "fi": "Finnish", "pl": "Polish", "cs": "Czech",
-    "fil": "Filipino", "tl": "Filipino", "fa": "Persian", "el": "Greek",
-    "ro": "Romanian", "hu": "Hungarian", "mk": "Macedonian",
-}
-_QWEN_CANONICAL = {v.lower(): v for v in _ISO_TO_QWEN.values()}
-
-
-def to_engine_language(code: str | None) -> str | None:
-    """上游 language（ISO-639-1 码 / 规范名 / 带地区子标签）→ Qwen3-ASR 规范名。
-
-    None/空/未识别 → None（交引擎自动检测），避免非法 hint 击穿到引擎抛错。
-    """
-    if code is None:
-        return None
-    s = str(code).strip()
-    if not s:
-        return None
-    key = s.lower()
-    if key in _QWEN_CANONICAL:                          # 已是规范名（Chinese/english…）
-        return _QWEN_CANONICAL[key]
-    if key in _ISO_TO_QWEN:                             # 纯 ISO 码（zh/en…）
-        return _ISO_TO_QWEN[key]
-    primary = key.replace("_", "-").split("-", 1)[0]    # zh-CN / en_US → 取主标签
-    return _ISO_TO_QWEN.get(primary)
+# 语言码归一已下沉到中立工具层 app/utils/language.py（原生实时/离线端点同样复用），
+# 此处 re-export 保持兼容层调用点与单测的导入路径不变。
+from app.utils.language import to_engine_language  # noqa: F401
 
 
 # v2 内部任务状态 → DashScope task_status（design §7.2）
